+++ elements/layercontrol/stories/expanded-layers.js
@@ -6,18 +6,13 @@ import {
   STORIES_MAP_STYLE,
 } from "../src/enums/stories";
 
-export const ExpandedLayers = {
-  args: {},
-  render: () => html`
-    <div style="display: flex">
-      <eox-layercontrol
-        for="eox-map#expanded"
-        .style=${STORIES_LAYERCONTROL_STYLE}
-      ></eox-layercontrol>
-      <eox-map
-        id="expanded"
-        .style=${STORIES_MAP_STYLE}
-        .layers=${[
+export const ExpandedLayersStory = {
+  args: {
+    for: "eox-map#expanded",
+    storyAdditionalComponents: {
+      "eox-map": {
+        style: STORIES_MAP_STYLE,
+        layers: [
           STORIES_LAYER_TERRAIN_LIGHT,
           {
             type: "Group",
@@ -27,11 +22,25 @@ export const ExpandedLayers = {
             },
             layers: [{ ...STORIES_LAYER_EOX_CLOUDLESS_2021, visible: false }],
           },
-        ]}
-      >
-      </eox-map>
+        ],
+        id: "expanded",
+      },
+    },
+    style: STORIES_LAYERCONTROL_STYLE,
+  },
+  render: (args) => html`
+    <div style="display: flex">
+      <eox-layercontrol
+        .for=${args.for}
+        .style=${args.style}
+      ></eox-layercontrol>
+      <eox-map
+        id=${args.storyAdditionalComponents["eox-map"].id}
+        .style=${args.storyAdditionalComponents["eox-map"].style}
+        .layers=${args.storyAdditionalComponents["eox-map"].layers}
+      ></eox-map>
     </div>
   `,
 };
 
-export default ExpandedLayers;
+export default ExpandedLayersStory;