+++ elements/storytelling/src/helpers/render-html-string.js
@@ -427,7 +427,7 @@ export function parseNavWithAddSection(
           ${nav
             .map(({ id, title, state }) =>
               state
-                ? `<li class="nav-${id}"><a href="#${id}"><small class="truncate">${title}</small></a></li>`
+                ? `<li class="nav-${id} center-align"><a href="#${id}"><small class="truncate">${title}</small></a></li>`
                 : "",
             )
             .join("")}