+++ elements/timecontrol/stories/disabled-buttons.js
@@ -7,15 +7,17 @@ export const DisabledButtons = {
     for: "eox-map#disabled-play",
     play: false,
     navigation: false,
+    storyAdditionalComponents: {
+      "eox-map": {
+        id: "disabled-play",
+        style: "width: 400px; height: 300px;",
+        zoom: DEFAULT_ARGS.zoom,
+        center: DEFAULT_ARGS.center,
+        layers: DEFAULT_ARGS.layers,
+      },
+    },
   },
   render: (args) => html`
-    <eox-map
-      id="disabled-play"
-      style="width: 400px; height: 300px;"
-      .zoom=${args.zoom}
-      .center=${args.center}
-      .layers=${args.layers}
-    ></eox-map>
     <eox-timecontrol
       .for=${args.for}
       .layer=${args.layer}
@@ -24,8 +26,14 @@ export const DisabledButtons = {
       .navigation=${args.navigation}
       .slider=${args.slider}
       .play=${args.play}
-      style="margin-top: 8px"
     ></eox-timecontrol>
+    <eox-map
+      id="${args.storyAdditionalComponents["eox-map"].id}"
+      style="${args.storyAdditionalComponents["eox-map"].style}"
+      .zoom=${args.storyAdditionalComponents["eox-map"].zoom}
+      .center=${args.storyAdditionalComponents["eox-map"].center}
+      .layers=${args.storyAdditionalComponents["eox-map"].layers}
+    ></eox-map>
   `,
 };